feat: Introduce pixels shifting of UV mapping
Without any shifting, if we see our hands we should see that there is
a huge gap to the left of fingers in the depth mask from the depth image
therefore we have to shift the fingers to the middle of this depth mask.
But take note that this shift distance is different at different depth
distances such as at 1 metre away from the camera, pixels are found aligned
in the depth mask therefore requires no shifting. Thus this shifting is
adaptive dependent on each pixel's depth value.
This is implemented by trial and error. Take note that the coordinates of points are calculated by u and v values ranging from 0.0 to 1.0. It is seen that at value=300 , that is, 30cm away from the camera, the shift amount of pixels in the u direction should be 0.04.
u     = 0.04 , 0.01 , 0.005 , 0.0025 , 0.00125
value = 300  , 450  , 600   , 750    , 900
So we had to calculate/make some function that takes input value and outputs
the right/close-to u-value. It took some hacking and trial and error, to produce
function u = 0.04 / 2.0^(((value/250.0)-1.0)/0.5)
Though u is not exactly the values above in the table, after trying it in the
application it seems reasonable.