fix: Change default values of Unity scenes; Clean PDepth.cs
Changes were made to complete the two features tested out before.
They are the adaptive shifting of pixels from the UV map to the
left based on pixel's depth distance from the camera. The second
feature is fixing the distance of how far in front should the
particle system + surroundings be rendered in front of the camera
within Unity. With the rendering of the particle system + surroundings
done in a single image, if we position it 95 world units in front
of the camera within Unity, it will have the same real-world
perspective when looking without wearing the Oculus Rift.